refactor: unify WalStorageConfig
